Improved: Split the color_565 module into multiple sub modules based on feature.
