+++ projects/dxt-lossless-transform-common/src/color_565/decorrelate.rs
@@ -1,0 +1,365 @@
+use super::Color565;
+use derive_enum_all_values::AllValues;
+
+/// Represents a function variant for decoration/recorrelation operations
+#[derive(Debug, Clone, Copy, PartialEq, Eq, AllValues, Hash)]
+pub enum YCoCgVariant {
+    /// Variant 1: Usually compresses best
+    Variant1,
+    /// Variant 2: Faster recorrelate (marginally) for compression speed
+    Variant2,
+    /// Variant 3: Sometimes better than variant 2
+    Variant3,
+    /// None: No transformation (original RGB values preserved)
+    None,
+}
+
+impl Color565 {
+    /// Transforms RGB color to YCoCg-R (reversible YCoCg) color space.
+    ///
+    /// YCoCg-R is a lifting-based variation of YCoCg that offers perfect reversibility.
+    /// This implementation treats all channels as 5-bit values for consistency.
+    ///
+    /// The YCoCg-R transformation follows these steps:
+    /// 1. Co = R - B
+    /// 2. t = B + (Co >> 1)
+    /// 3. Cg = G - t
+    /// 4. Y = t + (Cg >> 1)
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let mut color = Color565::from_rgb(255, 128, 64);
+    /// color.decorrelate_ycocg_r_var1();
+    /// // Color is now in YCoCg-R form
+    /// ```
+    #[inline]
+    pub fn decorrelate_ycocg_r_var1(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract RGB components
+        let r = (self.value >> 11) & 0x1F; // 5 bits for red
+        let g = (self.value >> 6) & 0x1F; // 5 top bits for green (ignoring bottom 1 bit)
+        let g_low = (self.value >> 5) & 0x1; // leftover bit for green
+        let b = self.value & 0x1F; // 5 bits for blue
+
+        // Apply YCoCg-R forward transform (all operations in 5-bit space)
+        // Step 1: Co = R - B
+        let co = (r as i16 - b as i16) & 0x1F;
+
+        // Step 2: t = B + (Co >> 1)
+        let t = (b as i16 + (co >> 1)) & 0x1F;
+
+        // Step 3: Cg = G - t
+        let cg = (g as i16 - t) & 0x1F;
+
+        // Step 4: Y = t + (Cg >> 1)
+        let y = (t + (cg >> 1)) & 0x1F;
+
+        // Pack into Color565 format:
+        // - Y (5 bits) in red position
+        // - Co (5 bits) in green position (shifted to use upper 5 bits of the 6-bit field)
+        // - Cg (5 bits) in blue position
+        Color565::from_raw(((y as u16) << 11) | ((co as u16) << 6) | (g_low << 5) | (cg as u16))
+    }
+
+    /// [Variant 1: Usually compresses best]
+    /// Transforms color from YCoCg-R back to RGB color space.
+    ///
+    /// This is the inverse of the decorrelation operation, following these steps:
+    /// 1. t = Y - (Cg >> 1)
+    /// 2. G = Cg + t
+    /// 3. B = t - (Co >> 1)
+    /// 4. R = B + Co
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let original = Color565::from_rgb(255, 128, 64);
+    /// let mut decorrelated = original;
+    /// decorrelated.decorrelate_ycocg_r_var1();
+    ///
+    /// // Now recorrelate it back to original
+    /// decorrelated.recorrelate_ycocg_r_var1();
+    /// assert_eq!(decorrelated.raw_value(), original.raw_value());
+    /// ```
+    #[inline]
+    pub fn recorrelate_ycocg_r_var1(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract YCoCg-R components
+        let y = (self.value >> 11) & 0x1F; // 5 bits (Y in red position)
+        let co = (self.value >> 6) & 0x1F; // 5 bits (Co in upper 5 bits of green position)
+        let g_low = (self.value >> 5) & 0x1; // Extract the preserved low bit of green
+        let cg = self.value & 0x1F; // 5 bits (Cg in blue position)
+
+        // Apply YCoCg-R inverse transform (all operations in 5-bit space)
+        // Step 1: t = Y - (Cg >> 1)
+        let t = (y as i16 - ((cg as i16) >> 1)) & 0x1F;
+
+        // Step 2: G = Cg + t
+        let g = (cg as i16 + t) & 0x1F;
+
+        // Step 3: B = t - (Co >> 1)
+        let b = (t - ((co as i16) >> 1)) & 0x1F;
+
+        // Step 4: R = B + Co
+        let r = (b + co as i16) & 0x1F;
+
+        // Pack back into RGB565 format, preserving the original g_low bit
+        Color565::from_raw(((r as u16) << 11) | ((g as u16) << 6) | (g_low << 5) | (b as u16))
+    }
+
+    /// [Variant 2: Faster recorrelate (marginally) for compression speed.]
+    /// Transforms RGB color to YCoCg-R (reversible YCoCg) color space.
+    ///
+    /// YCoCg-R is a lifting-based variation of YCoCg that offers perfect reversibility.
+    /// This implementation treats all channels as 5-bit values for consistency.
+    ///
+    /// The YCoCg-R transformation follows these steps:
+    /// 1. Co = R - B
+    /// 2. t = B + (Co >> 1)
+    /// 3. Cg = G - t
+    /// 4. Y = t + (Cg >> 1)
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let mut color = Color565::from_rgb(255, 128, 64);
+    /// color.decorrelate_ycocg_r_var2();
+    /// // Color is now in YCoCg-R form
+    /// ```
+    #[inline]
+    pub fn decorrelate_ycocg_r_var2(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract RGB components
+        let r = (self.value >> 11) & 0x1F; // 5 bits for red
+        let g = (self.value >> 6) & 0x1F; // 5 top bits for green (ignoring bottom 1 bit)
+        let g_low = (self.value >> 5) & 0x1; // leftover bit for green
+        let b = self.value & 0x1F; // 5 bits for blue
+
+        // Apply YCoCg-R forward transform (all operations in 5-bit space)
+        // Step 1: Co = R - B
+        let co = (r as i16 - b as i16) & 0x1F;
+
+        // Step 2: t = B + (Co >> 1)
+        let t = (b as i16 + (co >> 1)) & 0x1F;
+
+        // Step 3: Cg = G - t
+        let cg = (g as i16 - t) & 0x1F;
+
+        // Step 4: Y = t + (Cg >> 1)
+        let y = (t + (cg >> 1)) & 0x1F;
+
+        // Pack into Color565 format:
+        // - Y (5 bits) in red position
+        // - Co (5 bits) in green position (shifted to use upper 5 bits of the 6-bit field)
+        // - Cg (5 bits) in blue position
+        Color565::from_raw((g_low << 15) | ((y as u16) << 10) | ((co as u16) << 5) | (cg as u16))
+        // Note: Marginal speed improvement on recorrelate by placing low bit in the top.
+    }
+
+    /// [Variant 3: Sometimes better than variant 2.]
+    /// Transforms color from YCoCg-R back to RGB color space.
+    ///
+    /// This is the inverse of the decorrelation operation, following these steps:
+    /// 1. t = Y - (Cg >> 1)
+    /// 2. G = Cg + t
+    /// 3. B = t - (Co >> 1)
+    /// 4. R = B + Co
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let original = Color565::from_rgb(255, 128, 64);
+    /// let mut decorrelated = original;
+    /// decorrelated.decorrelate_ycocg_r_var2();
+    ///
+    /// // Now recorrelate it back to original
+    /// decorrelated.recorrelate_ycocg_r_var2();
+    /// assert_eq!(decorrelated.raw_value(), original.raw_value());
+    /// ```
+    #[inline]
+    pub fn recorrelate_ycocg_r_var2(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract YCoCg-R components
+        let g_low = self.value >> 15; // Extract the preserved low bit of green
+        let y = (self.value >> 10) & 0x1F; // 5 bits (Y in red position)
+        let co = (self.value >> 5) & 0x1F; // 5 bits (Co in upper 5 bits of green position)
+        let cg = self.value & 0x1F; // 5 bits (Cg in blue position)
+
+        // Apply YCoCg-R inverse transform (all operations in 5-bit space)
+        // Step 1: t = Y - (Cg >> 1)
+        let t = (y as i16 - ((cg as i16) >> 1)) & 0x1F;
+
+        // Step 2: G = Cg + t
+        let g = (cg as i16 + t) & 0x1F;
+
+        // Step 3: B = t - (Co >> 1)
+        let b = (t - ((co as i16) >> 1)) & 0x1F;
+
+        // Step 4: R = B + Co
+        let r = (b + co as i16) & 0x1F;
+
+        // Pack back into RGB565 format, preserving the original g_low bit
+        Color565::from_raw(((r as u16) << 11) | ((g as u16) << 6) | (g_low << 5) | (b as u16))
+    }
+
+    /// Transforms RGB color to YCoCg-R (reversible YCoCg) color space.
+    ///
+    /// YCoCg-R is a lifting-based variation of YCoCg that offers perfect reversibility.
+    /// This implementation treats all channels as 5-bit values for consistency.
+    ///
+    /// The YCoCg-R transformation follows these steps:
+    /// 1. Co = R - B
+    /// 2. t = B + (Co >> 1)
+    /// 3. Cg = G - t
+    /// 4. Y = t + (Cg >> 1)
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let mut color = Color565::from_rgb(255, 128, 64);
+    /// color.decorrelate_ycocg_r_var3();
+    /// // Color is now in YCoCg-R form
+    /// ```
+    #[inline]
+    pub fn decorrelate_ycocg_r_var3(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract RGB components
+        let r = (self.value >> 11) & 0x1F; // 5 bits for red
+        let g = (self.value >> 6) & 0x1F; // 5 top bits for green (ignoring bottom 1 bit)
+        let g_low = (self.value >> 5) & 0x1; // leftover bit for green
+        let b = self.value & 0x1F; // 5 bits for blue
+
+        // Apply YCoCg-R forward transform (all operations in 5-bit space)
+        // Step 1: Co = R - B
+        let co = (r as i16 - b as i16) & 0x1F;
+
+        // Step 2: t = B + (Co >> 1)
+        let t = (b as i16 + (co >> 1)) & 0x1F;
+
+        // Step 3: Cg = G - t
+        let cg = (g as i16 - t) & 0x1F;
+
+        // Step 4: Y = t + (Cg >> 1)
+        let y = (t + (cg >> 1)) & 0x1F;
+
+        // Pack into Color565 format:
+        // - Y (5 bits) in red position
+        // - Co (5 bits) in green position (shifted to use upper 5 bits of the 6-bit field)
+        // - Cg (5 bits) in blue position
+        Color565::from_raw(((y as u16) << 11) | ((co as u16) << 6) | ((cg as u16) << 1) | g_low)
+    }
+
+    /// Transforms color from YCoCg-R back to RGB color space.
+    ///
+    /// This is the inverse of the decorrelation operation, following these steps:
+    /// 1. t = Y - (Cg >> 1)
+    /// 2. G = Cg + t
+    /// 3. B = t - (Co >> 1)
+    /// 4. R = B + Co
+    ///
+    /// This variation has the g_low bit placed as the lowermost bit.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::Color565;
+    ///
+    /// let original = Color565::from_rgb(255, 128, 64);
+    /// let mut decorrelated = original;
+    /// decorrelated.decorrelate_ycocg_r_var3();
+    ///
+    /// // Now recorrelate it back to original
+    /// decorrelated.recorrelate_ycocg_r_var3();
+    /// assert_eq!(decorrelated.raw_value(), original.raw_value());
+    /// ```
+    #[inline]
+    pub fn recorrelate_ycocg_r_var3(&self) -> Self {
+        // 0x1F == 0b11111
+        // Extract YCoCg-R components
+        let y = (self.value >> 11) & 0x1F; // 5 bits (Y in red position)
+        let co = (self.value >> 6) & 0x1F; // 5 bits (Co in upper 5 bits of green position)
+        let cg = (self.value >> 1) & 0x1F; // 5 bits (Cg in bits 1-5)
+        let g_low = self.value & 0x1; // Extract the preserved low bit of green (lowermost bit)
+
+        // Apply YCoCg-R inverse transform (all operations in 5-bit space)
+        // Step 1: t = Y - (Cg >> 1)
+        let t = (y as i16 - ((cg as i16) >> 1)) & 0x1F;
+
+        // Step 2: G = Cg + t
+        let g = (cg as i16 + t) & 0x1F;
+
+        // Step 3: B = t - (Co >> 1)
+        let b = (t - ((co as i16) >> 1)) & 0x1F;
+
+        // Step 4: R = B + Co
+        let r = (b + co as i16) & 0x1F;
+
+        // Pack back into RGB565 format, preserving the original g_low bit
+        Color565::from_raw(((r as u16) << 11) | ((g as u16) << 6) | (g_low << 5) | (b as u16))
+    }
+
+    /// Applies the specified decorrelation variant to a color
+    ///
+    /// Wrapper around the variant-specific decorrelate methods.
+    ///
+    /// # Parameters
+    ///
+    /// - `variant`: The [`YCoCgVariant`] to use
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::{Color565, YCoCgVariant};
+    ///
+    /// let mut color = Color565::from_rgb(255, 128, 64);
+    /// color.decorrelate_ycocg_r(YCoCgVariant::Variant1);
+    /// // Color is now in YCoCg-R form
+    /// ```
+    #[inline]
+    pub fn decorrelate_ycocg_r(&self, variant: YCoCgVariant) -> Self {
+        match variant {
+            YCoCgVariant::Variant1 => self.decorrelate_ycocg_r_var1(),
+            YCoCgVariant::Variant2 => self.decorrelate_ycocg_r_var2(),
+            YCoCgVariant::Variant3 => self.decorrelate_ycocg_r_var3(),
+            YCoCgVariant::None => *self, // No transformation
+        }
+    }
+
+    /// Applies the specified recorrelation variant to a color
+    ///
+    /// Wrapper around the variant-specific recorrelate methods.
+    ///
+    /// # Parameters
+    ///
+    /// - `variant`: The [`YCoCgVariant`] to use
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use dxt_lossless_transform_common::color_565::{Color565, YCoCgVariant};
+    ///
+    /// let mut color = Color565::from_rgb(255, 128, 64);
+    /// color.decorrelate_ycocg_r(YCoCgVariant::Variant1);
+    /// color.recorrelate_ycocg_r(YCoCgVariant::Variant1);
+    /// ```
+    #[inline]
+    pub fn recorrelate_ycocg_r(&self, variant: YCoCgVariant) -> Self {
+        match variant {
+            YCoCgVariant::Variant1 => self.recorrelate_ycocg_r_var1(),
+            YCoCgVariant::Variant2 => self.recorrelate_ycocg_r_var2(),
+            YCoCgVariant::Variant3 => self.recorrelate_ycocg_r_var3(),
+            YCoCgVariant::None => *self, // No transformation
+        }
+    }
+}
